+++ CHANGELOG.md
@@ -9,6 +9,9 @@ CHANGELOG
 - Remove `$getSelectFields` parameter from `Field::authorize()` [\#1250 / mfn](https://github.com/rebing/graphql-laravel/pull/1250)
   it has been non-functional since half a decade
 
+### Fixed
+- Fix `SelectFields` crashing when field types use callable [\#1252 / mfn](https://github.com/rebing/graphql-laravel/pull/1252)
+
 2026-03-21, 10.0.0-RC2
 ----------------------