+++ CHANGELOG.md
@@ -11,6 +11,7 @@ CHANGELOG
 
 ### Fixed
 - Fix `SelectFields` crashing when field types use callable [\#1252 / mfn](https://github.com/rebing/graphql-laravel/pull/1252)
+- Fix APQ middleware race condition (TOCTOU) [\#1253 / mfn](https://github.com/rebing/graphql-laravel/pull/1253)
 
 2026-03-21, 10.0.0-RC2
 ----------------------